Corrected bugs and now all the converters are in the same file converters.py
